auth

+++ ErkinStudy.Domain/Entities/UserLesson.cs
@@ -2,9 +2,9 @@
 
 namespace ErkinStudy.Domain.Entities
 {
-    public class UserLesson
+	public class UserLesson
     {
-        public int UserId { get; set; }
+	    public long UserId { get; set; }
         public long LessonId { get; set; }
         public virtual ApplicationUser User { get; set; }
         public virtual Lesson Lesson { get; set; }
